backend/requirements.txt
fastapi==0.115.11
uvicorn==0.34.0
pandas==2.2.3
matplotlib==3.7.2
tensorflow==2.19.0  # Optional if the backend also needs to load the model
numpy==1.26.4  # Required for handling predictions
pillow==9.3.0  # For handling image uploads
gdown
python-multipart  # for file uploads
python-dotenv